refactor: centralize deploy runtime dependency service handling
Move runtime contract stack application and contract-only dependency service resolution out of operations.ts and into the runtime contract layer, so deploy prepare no longer owns SQL dependency inference or Spring datasource connection env construction.

Keep dependency runtime signal matching behind shared dependency helpers and preserve DATASOURCE_URL detection while validating the deploy structured blocker and smoke flows.
