+++ src/core/deployment/operations.ts
@@ -50,7 +50,7 @@ import {
   resolveComposePath,
 } from "./runtime";
 import { writeDeploymentRepairRequest } from "./repair";
-import { deploymentRuntimeContractFromAac } from "./runtime-contract";
+import { applyRuntimeContractToStack, deploymentRuntimeContractFromAac } from "./runtime-contract";
 import { resolveDeploymentStrategy } from "./strategy";
 import { validateStartupLogs } from "./validate";
 import { discoverNodeWorkspacePackageJsonPaths, resolveDeploymentWorkspaceForApp } from "./workspace";
@@ -1805,273 +1805,6 @@ async function loadDeploymentRuntimeContractForLocator(
   } catch {
     return deploymentRuntimeContractFromAac(null, stack, null);
   }
-}
-
-function applyRuntimeContractToStack(
-  stack: DeploymentSpec["detectedStack"],
-  runtimeContract: DeploymentSpec["runtimeContract"],
-): DeploymentSpec["detectedStack"] {
-  const inferredKind = inferRuntimeContractStackKind(stack, runtimeContract);
-  const inferredPackageManager = inferRuntimeContractPackageManager(stack, runtimeContract, inferredKind);
-  return {
-    ...stack,
-    kind: inferredKind,
-    packageManager: inferredPackageManager,
-    framework: stack.framework ?? inferRuntimeContractFramework(runtimeContract, inferredKind),
-    buildCommand: runtimeContract.buildCommand ?? stack.buildCommand,
-    startCommand: deploymentStartCommand(stack, runtimeContract),
-    healthcheckPath: runtimeContract.healthPath ?? runtimeContract.previewPath ?? stack.healthcheckPath,
-    outputDirectory: runtimeContract.frontendOutputDir ?? stack.outputDirectory,
-    port: runtimeContract.port ?? stack.port,
-    services: runtimeContract.dependencyServicePolicy === "contract_only"
-      ? contractOnlyDependencyServices(runtimeContract, stack.services)
-      : mergeDependencyServices(stack.services, runtimeContract.dependencyServices),
-  };
-}
-
-function inferRuntimeContractStackKind(
-  stack: DeploymentSpec["detectedStack"],
-  runtimeContract: DeploymentSpec["runtimeContract"],
-): DeploymentSpec["detectedStack"]["kind"] {
-  if (stack.kind !== "unknown" || runtimeContract.source === "heuristic") {
-    return stack.kind;
-  }
-
-  const signals = runtimeContractSignals(runtimeContract);
-  if (/\b(node|npm|pnpm|yarn|bun|vite|next|react|express|fastify|hono|koa)\b/.test(signals)) {
-    return "node";
-  }
-  if (/\b(python|pip|poetry|uv|uvicorn|gunicorn|fastapi|flask|django)\b/.test(signals)) {
-    return "python";
-  }
-  if (/\b(go|golang)\b/.test(signals)) {
-    return "go";
-  }
-  if (/\b(java|maven|gradle|spring)\b/.test(signals)) {
-    return "java";
-  }
-  if (/\b(dotnet|aspnet|csharp|c#)\b/.test(signals)) {
-    return "dotnet";
-  }
-  if (/\b(php|composer|laravel|symfony)\b/.test(signals)) {
-    return "php";
-  }
-  if (/\b(ruby|bundler|bundle|rails|sinatra)\b/.test(signals)) {
-    return "ruby";
-  }
-  if (runtimeContract.frontendOutputDir && !runtimeContract.startCommand) {
-    return "static";
-  }
-  return stack.kind;
-}
-
-function inferRuntimeContractPackageManager(
-  stack: DeploymentSpec["detectedStack"],
-  runtimeContract: DeploymentSpec["runtimeContract"],
-  kind: DeploymentSpec["detectedStack"]["kind"],
-): DeploymentSpec["detectedStack"]["packageManager"] {
-  if (stack.packageManager || runtimeContract.source === "heuristic") {
-    return stack.packageManager;
-  }
-
-  const signals = runtimeContractSignals(runtimeContract);
-  if (kind === "node") {
-    if (/\bpnpm\b/.test(signals)) {
-      return "pnpm";
-    }
-    if (/\byarn\b/.test(signals)) {
-      return "yarn";
-    }
-    if (/\bbun\b/.test(signals)) {
-      return "bun";
-    }
-    return "npm";
-  }
-  if (kind === "python") {
-    if (/\bpoetry\b/.test(signals)) {
-      return "poetry";
-    }
-    if (/\buv\b/.test(signals)) {
-      return "uv";
-    }
-    return "pip";
-  }
-  if (kind === "java") {
-    return /\bgradle\b/.test(signals) ? "gradle" : "maven";
-  }
-  if (kind === "dotnet") {
-    return "dotnet";
-  }
-  if (kind === "go") {
-    return "go";
-  }
-  if (kind === "php") {
-    return "composer";
-  }
-  if (kind === "ruby") {
-    return "bundler";
-  }
-  return stack.packageManager;
-}
-
-function inferRuntimeContractFramework(
-  runtimeContract: DeploymentSpec["runtimeContract"],
-  kind: DeploymentSpec["detectedStack"]["kind"],
-): string | null {
-  if (runtimeContract.source === "heuristic") {
-    return null;
-  }
-  const signals = runtimeContractSignals(runtimeContract);
-  if (kind === "node") {
-    if (/\bvite\b/.test(signals)) {
-      return "vite";
-    }
-    if (/\bnext\b/.test(signals)) {
-      return "next";
-    }
-    if (/\b(express|fastify|hono|koa)\b/.test(signals)) {
-      return "node-server";
-    }
-    return "node";
-  }
-  return null;
-}
-
-function runtimeContractSignals(runtimeContract: DeploymentSpec["runtimeContract"]): string {
-  return [
-    runtimeContract.runtimeKind,
-    runtimeContract.buildCommand,
-    runtimeContract.startCommand,
-    runtimeContract.frontendOutputDir,
-    ...runtimeContract.environment.required,
-    ...runtimeContract.environment.optional,
-    ...runtimeContract.dependencyServices.flatMap((service) => [
-      service.kind,
-      service.serviceName,
-      service.image,
-      ...Object.keys(service.connectionEnv),
-      ...Object.values(service.connectionEnv),
-    ]),
-    ...runtimeContract.apiPaths,
-  ]
-    .filter((value): value is string => typeof value === "string")
-    .join("\n")
-    .toLowerCase();
-}
-
-function mergeDependencyServices(
-  detected: DeploymentSpec["detectedStack"]["services"],
-  declared: DeploymentSpec["detectedStack"]["services"],
-): DeploymentSpec["detectedStack"]["services"] {
-  const byKind = new Map<string, DeploymentSpec["detectedStack"]["services"][number]>();
-  for (const service of detected) {
-    byKind.set(service.kind, service);
-  }
-  for (const service of declared) {
-    byKind.set(service.kind, service);
-  }
-  return [...byKind.values()];
-}
-
-function contractOnlyDependencyServices(
-  runtimeContract: DeploymentSpec["runtimeContract"],
-  detected: DeploymentSpec["detectedStack"]["services"],
-): DeploymentSpec["detectedStack"]["services"] {
-  if (runtimeContract.dependencyServices.length > 0) {
-    return runtimeContract.dependencyServices;
-  }
-  if (!runtimeContractRequestsSqlDatabase(runtimeContract)) {
-    return [];
-  }
-  return detected
-    .filter((service) => ["postgres", "mysql"].includes(service.kind))
-    .map((service) => serviceWithRuntimeContractConnectionEnv(service, runtimeContract));
-}
-
-function runtimeContractRequestsSqlDatabase(runtimeContract: DeploymentSpec["runtimeContract"]): boolean {
-  const signals = [
-    runtimeContract.runtimeKind,
-    runtimeContract.buildCommand,
-    runtimeContract.startCommand,
-    ...runtimeContract.environment.required,
-    ...runtimeContract.environment.optional,
-    ...runtimeContract.apiPaths,
-  ]
-    .filter((value): value is string => typeof value === "string")
-    .join("\n")
-    .toLowerCase();
-  return /database_url|database_uri|spring_datasource|spring\.datasource|datasource_url|db_url|jdbc/.test(signals);
-}
-
-function serviceWithRuntimeContractConnectionEnv(
-  service: DeploymentSpec["detectedStack"]["services"][number],
-  runtimeContract: DeploymentSpec["runtimeContract"],
-): DeploymentSpec["detectedStack"]["services"][number] {
-  const requested = new Set([
-    ...runtimeContract.environment.required,
-    ...runtimeContract.environment.optional,
-  ]);
-  if (
-    !requested.has("SPRING_DATASOURCE_URL") &&
-    !requested.has("SPRING_DATASOURCE_USERNAME") &&
-    !requested.has("SPRING_DATASOURCE_PASSWORD")
-  ) {
-    return service;
-  }
-
-  if (service.kind === "postgres") {
-    return {
-      ...service,
-      connectionEnv: {
-        ...service.connectionEnv,
-        SPRING_DATASOURCE_URL: "jdbc:postgresql://postgres:5432/loom",
-        SPRING_DATASOURCE_USERNAME: "loom",
-        SPRING_DATASOURCE_PASSWORD: "loom",
-      },
-    };
-  }
-  if (service.kind === "mysql") {
-    return {
-      ...service,
-      connectionEnv: {
-        ...service.connectionEnv,
-        SPRING_DATASOURCE_URL: "jdbc:mysql://mysql:3306/loom",
-        SPRING_DATASOURCE_USERNAME: "loom",
-        SPRING_DATASOURCE_PASSWORD: "loom",
-      },
-    };
-  }
-  return service;
-}
-
-function deploymentStartCommand(
-  stack: DeploymentSpec["detectedStack"],
-  runtimeContract: DeploymentSpec["runtimeContract"],
-): string | null {
-  const contractCommand = runtimeContract.startCommand;
-  const detectedCommand = stack.startCommand;
-  const targetPort = runtimeContract.port ?? stack.port;
-
-  if (contractCommand && detectedCommand && isLongLivedDevCommand(contractCommand) && !isLongLivedDevCommand(detectedCommand)) {
-    return ensureStartCommandPort(detectedCommand, stack, targetPort);
-  }
-
-  return ensureStartCommandPort(contractCommand ?? detectedCommand, stack, targetPort);
-}
-
-function isLongLivedDevCommand(command: string): boolean {
-  return /\b(dev|watch)\b/i.test(command);
-}
-
-function ensureStartCommandPort(
-  command: string | null,
-  stack: DeploymentSpec["detectedStack"],
-  port: number,
-): string | null {
-  if (!command || stack.framework !== "vite" || !/\bpreview\b/i.test(command) || /\s--port(?:=|\s)/.test(command)) {
-    return command;
-  }
-  return `${command} --port ${port}`;
 }
 
 function applyRuntimeContractHealthDefaults(
